+++ src/pages/about/AboutPage.tsx
@@ -1,5 +1,5 @@
-import About from '../components/about/About';
-import GlobalStyles from '../styles/GlobalStyle';
+import About from '../../components/about/About';
+import GlobalStyles from '../../styles/GlobalStyle';
 //import '../components/LoadScript';
 
 const AboutPage = () => {
