+++ src/pages/form/login/LoginPage.tsx
@@ -1,4 +1,4 @@
-import { Login } from '../components/login/Login';
+import { Login } from '../../../components/login/Login';
 //import '../components/LoadScript';
 
 const LoginPage = () => {
