+++ src/pages/form/signUp/SignUp.tsx
@@ -1,7 +1,6 @@
-import '../styles/signUp.css';
-import Sform from '../components/signUp/Sform';
-import SignupModal from '../components/login/mobile/SignupModal';
-import MMoreInfo from '../components/login/mobile/MMoreInfo';
+import MMoreInfo from '../../../components/login/mobile/MMoreInfo';
+import Sform from '../../../components/signUp/Sform';
+import '../../../styles/signUp.css';
 //import '../components/LoadScript';
 
 const SignUp = () => {
