+++ src/pages/mypage/modify/UserModify.tsx
@@ -1,4 +1,4 @@
-import UserInfoModify from '../components/mypage/UserInfoModify';
+import UserInfoModify from '../../../components/mypage/UserInfoModify';
 
 const UserModify = () => {
     return (
