+++ src/pages/recruit/RecruitPage.tsx
@@ -1,6 +1,3 @@
-import React from 'react';
-import UnivRecruit from '../components/univrecruit/UnivRecruit';
-import Recruit from '../components/recruit/Recruit';
 import { Outlet } from 'react-router-dom';
 //import '../components/LoadScript';
