+++ src/pages/univ/UnivPage.tsx
@@ -1,6 +1,6 @@
 // 참여 대학 페이지
-import UnivTab from '../components/univ/UnivTab';
-import UnivHeader from '../components/univ/UnivHeader';
+import UnivTab from '../../components/univ/UnivTab';
+import UnivHeader from '../../components/univ/UnivHeader';
 import { Suspense } from 'react';
 
 const UnivPage = () => {
