+++ src/pages/userpage/Userpage.tsx
@@ -1,7 +1,7 @@
 import { useParams } from 'react-router-dom';
 
-import UserInfomation from '../components/mypage/UserInfomation';
-import UserPost from '../components/mypage/UserPost';
+import UserInfomation from '../../components/mypage/UserInfomation';
+import UserPost from '../../components/mypage/UserPost';
 
 const Userpage = () => {
     let { user_id } = useParams();
