+++ tsconfig.json
@@ -16,5 +16,5 @@
         "noEmit": true,
         "jsx": "react-jsx"
     },
-    "include": ["src", "custom.d.ts"]
+    "include": ["src", "custom.d.ts", "src/pages/about/AboutPage.tsx"]
 }
